--- src/App.scss
@@ -1,113 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.title{
-  font-family: 'Unbounded';
-  font-size: 2rem;
-}
-
-.header {
-  padding: 2rem;
-  border-bottom: 0.1rem solid #ccc;
-  position: fixed;
-  background: white;
-  width: 100%;
-  top: 0;
-  z-index: 9999;
-}
-
-.body {
-  display: block;
-  position: absolute;
-  top: 11rem;
-  scroll-behavior: auto;
-  padding: 2rem;
-}
-
-.label {
-  font-weight: bold;
-}
-
-.releaseTitle {
-  font-family: 'Unbounded';
-  font-weight: bold;
-  font-size: 1.2rem;
-}
-
-.pr_separator {
-  margin: 1rem 0;
-}
-
-.prs_title_open {
-  font-weight: bold;
-  font-size: 1rem;
-  background-color: #ccc;
-  color: #505050;
-  padding: 1rem;
-  border-top-right-radius: 0.5rem;
-  border-top-left-radius: 0.5rem;
-  border: 0.1rem solid #ccc;
-  cursor: pointer;
-}
-
-.prs_title_loading {
-  display:flex;
-  flex-direction: row;
-  font-weight: bold;
-  font-size: 1rem;
-  background-color: #ccc;
-  color: #505050;
-  padding: 1rem;
-  border-radius: 0.5rem;
-  border-bottom-left-radius: 0.5rem;
-  border-bottom-right-radius: 0.5rem;
-  border: 0.1rem solid #ccc;
-}
-
-.loading{
-  margin-left: 1rem;
-}
-
-.prs_title_close {
-  font-weight: bold;
-  font-size: 1rem;
-  background-color: #ccc;
-  color: #505050;
-  padding: 1rem;
-  border-radius: 0.5rem;
-  border-bottom-left-radius: 0.5rem;
-  border-bottom-right-radius: 0.5rem;
-  border: 0.1rem solid #ccc;
-  cursor: pointer;
-}
-
-.prs {
-  background-color: #ccc;
-  padding: 1rem;
-  border-bottom-left-radius: 0.5rem;
-  border-bottom-right-radius: 0.5rem;
-  border: 0.1rem solid #ccc;
-}
-
-.version {
-  padding-left: 1rem;
-  font-size: 0.6rem;
-}
-
-.inputWrapper{
-  margin-top: 2rem;
-  width: 50%;
-  display: flex;
-  flex-direction: column;
-}
-
-.searchInput {
-  width: 100%;
-  padding: 12px 20px;
-  margin: 8px 0;
-  display: inline-block;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-  box-sizing: border-box;
-}
