+++ src/ThemeSelector.scss
@@ -1,0 +1,7 @@
+.themeSelector {
+  z-index: 999 !important;
+  position: fixed;
+  display: block;
+  right: 2rem;
+  top: 1rem;
+}
